# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO API",
        "description": "Todo API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/",
            "description": "LIVE server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "example@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "register",
                "description": "Register",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Example"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "example@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/checklists": {
            "get": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Display list of checklists",
                "description": "Get checklists",
                "operationId": "dacf4d6ce3960f869f770bb28cb12532",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Store checklists",
                "description": "Store checklists",
                "operationId": "ab63946a17fae9a4373c90fabb59630a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Checklist 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/checklists/{id}": {
            "delete": {
                "tags": [
                    "Checklist"
                ],
                "summary": "Show checklists",
                "description": "Get checklists",
                "operationId": "fb93efa792cf08fd33835d8e3e0b80e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/checklists/{id}/item": {
            "get": {
                "tags": [
                    "Checklist Item"
                ],
                "summary": "OA checklist items",
                "description": "Get checklist items",
                "operationId": "01170091eefbfdcf84ac66e95c97181e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Checklist Item"
                ],
                "summary": "Show checklists",
                "description": "Get checklists",
                "operationId": "cfd8c2c123f7f240ae8e5787b680f8e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Checklist Item"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/checklists/{id}/item/{itemId}": {
            "get": {
                "tags": [
                    "Checklist Item"
                ],
                "summary": "OA checklist items",
                "description": "Get checklist items",
                "operationId": "2225a1580ed6aeb56013829a28cc4918",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Checklist Item"
                ],
                "summary": "Show checklists",
                "description": "Get checklists",
                "operationId": "173708b392137cfa6773b463ed0aa04e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Checklist 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Checklist Item"
                ],
                "summary": "Show checklists",
                "description": "Get checklists",
                "operationId": "94ae27b81ebd0f96b85a81b9370ccd09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Checklist 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/checklists/{id}/item/rename/{itemId}": {
            "put": {
                "tags": [
                    "Checklist Item"
                ],
                "summary": "Show checklists",
                "description": "Get checklists",
                "operationId": "580079a80aeb5a17a81e434f46e57880",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Checklist 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}